+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 ### New Feature
 
 - Added component `DiagramNode` for CA Learning graph
+- Added styling for `DiagramNode` when selected in CA Learning graph
+- Added UI for the test yourself page
+- Added component `ExerciseBoard` for the exercise page
+- Added styling and subcomponents for `ExerciseBoard`
+-   Added component `ComponentBoard` and its styling
+-   Added component `ComponentPieces` and its styling
+-   Added component `ComponentPiecesDroppableArea` and its styling
+-   Added component `ExerciseBoardSidebar` and its styling
+-   Added component `ExerciseBoardSubheader` and its styling
 
 ### Bug Fix